services/ingestion/requirements.txt
psycopg2-binary
numpy
pyarrow
boto3
